feat: change homepage and use the b variation homepage (#1063)
