+++ src/pages/home/_vertical-carousel.js
